app 폴더 생성 후 기본 element 생성 함수들은 index.js에, 로직 함수들은 logic.js에 옮기기
