+++ src/basic/app/logic.js
@@ -1,0 +1,150 @@
+import { ElementIds } from "../../constants.js";
+import { prodList } from "../store/prodList.js";
+import { calcCart } from "../calcCart.js";
+import { lastSel, updateLastSelValue } from "../store/lastSel.js";
+
+function isSoldOut(quantity) {
+  return quantity === 0;
+}
+
+function createOptionFromItem(item) {
+  const opt = document.createElement("option");
+  opt.value = item.id;
+  opt.textContent = item.name + " - " + item.val + "원";
+  opt.disabled = isSoldOut(item.q);
+  return opt;
+}
+
+export function updateSelOpts(selElem) {
+  const sel = selElem ?? document.getElementById(ElementIds.SEL);
+
+  sel.innerHTML = "";
+
+  prodList.forEach((item) => {
+    const opt = createOptionFromItem(item);
+    sel.appendChild(opt);
+  });
+}
+
+export function handleClickAddBtn() {
+  const sel = document.getElementById(ElementIds.SEL);
+  const cartDisp = document.getElementById(ElementIds.CART_DISP);
+  var selItem = sel.value;
+  var itemToAdd = prodList.find(function (p) {
+    return p.id === selItem;
+  });
+  if (itemToAdd && itemToAdd.q > 0) {
+    var item = document.getElementById(itemToAdd.id);
+    if (item) {
+      var newQty =
+        parseInt(item.querySelector("span").textContent.split("x ")[1]) + 1;
+      if (newQty <= itemToAdd.q) {
+        item.querySelector("span").textContent =
+          itemToAdd.name + " - " + itemToAdd.val + "원 x " + newQty;
+        itemToAdd.q--;
+      } else {
+        alert("재고가 부족합니다.");
+      }
+    } else {
+      var newItem = document.createElement("div");
+      newItem.id = itemToAdd.id;
+      newItem.className = "flex justify-between items-center mb-2";
+      newItem.innerHTML =
+        "<span>" +
+        itemToAdd.name +
+        " - " +
+        itemToAdd.val +
+        "원 x 1</span><div>" +
+        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
+        itemToAdd.id +
+        '" data-change="-1">-</button>' +
+        '<button class="quantity-change bg-blue-500 text-white px-2 py-1 rounded mr-1" data-product-id="' +
+        itemToAdd.id +
+        '" data-change="1">+</button>' +
+        '<button class="remove-item bg-red-500 text-white px-2 py-1 rounded" data-product-id="' +
+        itemToAdd.id +
+        '">삭제</button></div>';
+      cartDisp.appendChild(newItem);
+      itemToAdd.q--;
+    }
+    calcCart();
+    updateLastSelValue(selItem);
+  }
+}
+
+export function handleClickCartDisp(event) {
+  var tgt = event.target;
+  if (
+    tgt.classList.contains("quantity-change") ||
+    tgt.classList.contains("remove-item")
+  ) {
+    var prodId = tgt.dataset.productId;
+    var itemElem = document.getElementById(prodId);
+    var prod = prodList.find(function (p) {
+      return p.id === prodId;
+    });
+    if (tgt.classList.contains("quantity-change")) {
+      var qtyChange = parseInt(tgt.dataset.change);
+      var newQty =
+        parseInt(itemElem.querySelector("span").textContent.split("x ")[1]) +
+        qtyChange;
+      if (
+        newQty > 0 &&
+        newQty <=
+          prod.q +
+            parseInt(itemElem.querySelector("span").textContent.split("x ")[1])
+      ) {
+        itemElem.querySelector("span").textContent =
+          itemElem.querySelector("span").textContent.split("x ")[0] +
+          "x " +
+          newQty;
+        prod.q -= qtyChange;
+      } else if (newQty <= 0) {
+        itemElem.remove();
+        prod.q -= qtyChange;
+      } else {
+        alert("재고가 부족합니다.");
+      }
+    } else if (tgt.classList.contains("remove-item")) {
+      var remQty = parseInt(
+        itemElem.querySelector("span").textContent.split("x ")[1],
+      );
+      prod.q += remQty;
+      itemElem.remove();
+    }
+
+    calcCart();
+  }
+}
+
+export function setSaleAlert() {
+  setTimeout(function () {
+    setInterval(function () {
+      var luckyItem = prodList[Math.floor(Math.random() * prodList.length)];
+      if (Math.random() < 0.3 && luckyItem.q > 0) {
+        luckyItem.val = Math.round(luckyItem.val * 0.8);
+        alert("번개세일! " + luckyItem.name + "이(가) 20% 할인 중입니다!");
+        updateSelOpts();
+      }
+    }, 30000);
+  }, Math.random() * 10000);
+}
+
+export function setSuggestionAlert() {
+  setTimeout(function () {
+    setInterval(function () {
+      if (lastSel) {
+        var suggest = prodList.find(function (item) {
+          return item.id !== lastSel && item.q > 0;
+        });
+        if (suggest) {
+          alert(
+            suggest.name + "은(는) 어떠세요? 지금 구매하시면 5% 추가 할인!",
+          );
+          suggest.val = Math.round(suggest.val * 0.95);
+          updateSelOpts();
+        }
+      }
+    }, 60000);
+  }, Math.random() * 20000);
+}
